requirements.txt
# Discord Bot Framework
discord.py>=2.3.0

# Environment Configuration
python-dotenv>=1.0.0

# HTTP Client for Pterodactyl API
aiohttp>=3.9.0

# Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Code Quality
black>=23.0.0
flake8>=6.1.0
